pyproject.toml
[tool.poetry]
name = "cryptbuddy"
version = "0.0.5"
description = ""
authors = ["Kush Patel <kush@kush.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.9.0"}
PyNaCl = "^1.5.0"
password-strength = "^0.0.3.post2"
appdirs = "^1.4.4"
pytest = "^7.3.1"

[tool.poetry.scripts]
cryptbuddy = "cryptbuddy.main:app"
crypt = "cryptbuddy.main:app"
cb = "cryptbuddy.main:app"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
pdoc = "^14.1.0"
setuptools = "^68.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
